Remove require cycles on Annotation and maps/index
