+++ javascript/index.js
@@ -29,6 +29,7 @@ import locationManager from './modules/location/locationManager';
 import offlineManager from './modules/offline/offlineManager';
 import snapshotManager from './modules/snapshot/snapshotManager';
 import MarkerView from './components/MarkerView';
+import Animated from './utils/animated/Animated';
 import AnimatedMapPoint from './utils/animated/AnimatedPoint';
 import AnimatedShape from './utils/animated/AnimatedShape';
 import AnimatedCoordinatesArray from './utils/animated/AnimatedCoordinatesArray';
@@ -98,36 +99,15 @@ MapboxGL.locationManager = locationManager;
 MapboxGL.offlineManager = offlineManager;
 MapboxGL.snapshotManager = snapshotManager;
 
+// animated
+MapboxGL.Animated = Animated;
+
 // utils
 MapboxGL.AnimatedPoint = AnimatedMapPoint;
 MapboxGL.AnimatedCoordinatesArray = AnimatedCoordinatesArray;
 MapboxGL.AnimatedExtractCoordinateFromArray = AnimatedExtractCoordinateFromArray;
 MapboxGL.AnimatedRouteCoordinatesArray = AnimatedRouteCoordinatesArray;
 MapboxGL.AnimatedShape = AnimatedShape;
-
-// animated
-const Animated = {
-  // sources
-  ShapeSource: RNAnimated.createAnimatedComponent(ShapeSource),
-  ImageSource: RNAnimated.createAnimatedComponent(ImageSource),
-
-  // layers
-  FillLayer: RNAnimated.createAnimatedComponent(FillLayer),
-  FillExtrusionLayer: RNAnimated.createAnimatedComponent(FillExtrusionLayer),
-  LineLayer: RNAnimated.createAnimatedComponent(LineLayer),
-  CircleLayer: RNAnimated.createAnimatedComponent(CircleLayer),
-  SymbolLayer: RNAnimated.createAnimatedComponent(SymbolLayer),
-  RasterLayer: RNAnimated.createAnimatedComponent(RasterLayer),
-  BackgroundLayer: RNAnimated.createAnimatedComponent(BackgroundLayer),
-
-  // values
-  CoordinatesArray: AnimatedCoordinatesArray,
-  RouteCoordinatesArray: AnimatedRouteCoordinatesArray,
-  Shape: AnimatedShape,
-  ExtractCoordinateFromArray : AnimatedExtractCoordinateFromArray, 
-};
-
-MapboxGL.Animated = Animated;
 
 const {LineJoin} = MapboxGL;
